Move challenge result to legacy package
